Server/requirements.txt
streamlit
Pillow
fpdf
langchain
langchain-groq
requests
python-dotenv
python-docx
agno
duckduckgo-search
matplotlib
torch
transformers
sentencepiece
scikit-image
google-genai
hi-ml-multimodal
IPython
openai==0.28
diffusers
accelerate
safetensors
